pyproject.toml
[project]
name = "newspaper-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-documentai>=3.5.0",
    "google-cloud-documentai-toolbox>=0.14.2a0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.1",
    "matplotlib>=3.10.3",
    "numpy>=1.26.4,<2.0",
    "opencc>=1.1.9",
    "opencv-python>=4.11.0.86",
    "pillow>=10.4.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "scipy>=1.16.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["newspaper_ocr"]
